+++ src/mbtaclient/handlers/trips_handler.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from typing import Optional
 import logging
 
-from ..trip_stop import StopType
+from ..stop import StopType
 
 from ..client.mbta_client import MBTAClient
 from ..handlers.base_handler import MBTABaseHandler
